+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - Add uninstall script to switch from winesapOS back to upstream Arch Linux.
 - Add support for Mac drivers on Linux LTS 6.1.
 - Add support for installing the GNOME desktop environment.
+- Add the AppImagePool package manager.
 
 ## 3.2.1 - The SteamOS 3.4 Update